refactor: aop로 exception handler에서 slack 의존성 제거
